change: moved to src/services
* updated imports to import from src/service
* added index.ts to UniversimeApi
