+++ src/services/UniversimeApi/index.ts
@@ -1,0 +1,4 @@
+import { UniversimeApi } from "./UniversimeApi";
+
+export default UniversimeApi;
+export * from "./UniversimeApi";
